tested pypi setup on both macos and windows10 refactor: moved utils.imgCrop to utils.imgProcessing refactor: split runReconstruction into two, one for arg parsing and another to workflow.reconstructBatch for processing
