+++ ReconstructOrder/compute/__init__.py
@@ -1,4 +1,8 @@
 # bchhun, {4/23/19}
 
 name = "compute"
-from .reconstruct import ImgReconstructor
+# from .reconstruct import ImgReconstructor
+
+from . import reconstruct
+
+__all__ = ["reconstruct"]
