+++ ReconstructOrder/utils/__init__.py
@@ -1,18 +1,26 @@
 name = "utils"
 
-from .ConfigReader import ConfigReader, Dataset, Processing, Plotting
+# from .ConfigReader import ConfigReader, Dataset, Processing, Plotting
+#
+# from .imgCrop import imcrop, toggle_selector, line_select_callback
+#
+# from .imgIO import GetSubDirName, FindDirContainPos, loadTiff, ParseFileList, \
+#     ParseTiffInput_old, parse_tiff_input, sort_pol_channels, exportImg, \
+#     process_position_list, process_timepoint_list, process_z_slice_list
+#
+# from .imgProcessing import ImgMin, ImgLimit, nanRobustBlur, histequal, imBitConvert, \
+#     imadjust, imadjustStack, imClip, linScale, removeBubbles
+#
+# from .mManagerIO import mManagerReader, PolAcquReader
+#
+# from .plotting import plotVectorField, PolColor, CompositeImg, \
+#     plot_recon_images, plot_stokes, plot_pol_imgs, plot_Polacquisition_imgs, plot_sub_images, \
+#     render_birefringence_imgs
 
-from .imgCrop import imcrop, toggle_selector, line_select_callback
+__all__ = ["ConfigReader",
+           "imgIO",
+           "imgProcessing",
+           "mManagerIO",
+           "plotting"]
 
-from .imgIO import GetSubDirName, FindDirContainPos, loadTiff, ParseFileList, \
-    ParseTiffInput_old, parse_tiff_input, sort_pol_channels, exportImg, \
-    process_position_list, process_timepoint_list, process_z_slice_list
-
-from .imgProcessing import ImgMin, ImgLimit, nanRobustBlur, histequal, imBitConvert, \
-    imadjust, imadjustStack, imClip, linScale, removeBubbles
-
-from .mManagerIO import mManagerReader, PolAcquReader
-
-from .plotting import plotVectorField, PolColor, CompositeImg, \
-    plot_recon_images, plot_stokes, plot_pol_imgs, plot_Polacquisition_imgs, plot_sub_images, \
-    render_birefringence_imgs
+from . import imgIO, imgProcessing, plotting, ConfigReader
