+++ ReconstructOrder/workflow/__init__.py
@@ -1,7 +1,13 @@
 name = "workflow"
 
-from .processImages import runReconstruction
+# from .processImages import runReconstruction
 
-from .multiDimProcess import create_metadata_object, parse_tiff_input, parse_bg_options, \
-    process_background, compute_flat_field, correct_flat_field, read_metadata, \
-    loopPos, loopT, loopZSm, loopZBg
+# from .multiDimProcess import create_metadata_object, parse_tiff_input, parse_bg_options, \
+#     process_background, compute_flat_field, correct_flat_field, read_metadata, \
+#     loopPos, loopT, loopZSm, loopZBg
+
+__all__ = ["multiDimProcess",
+           "reconstructBatch"]
+
+from . import multiDimProcess
+from .reconstructBatch import reconstructBatch
