+++ ReconstructOrder/workflow/multiDimProcess.py
@@ -1,10 +1,16 @@
 import os
 import numpy as np
+import matplotlib
+matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import cv2
 
-from ..utils import *
-from ..compute import ImgReconstructor
+from ..utils.imgIO import parse_tiff_input, exportImg
+from ..utils.imgProcessing import ImgMin, imBitConvert
+from ..utils.mManagerIO import mManagerReader, PolAcquReader
+from ..utils.plotting import plot_stokes, plot_pol_imgs, plot_Polacquisition_imgs, \
+    render_birefringence_imgs
+from ..compute.reconstruct import ImgReconstructor
 
 
 def create_metadata_object(data_path, config):
