+++ examples/example_configs/config_MouseBrainSlice1_workflow_test.yml
@@ -1,6 +1,6 @@
 dataset:
     data_dir: './example_data/TestData/rawData/2018_10_02_MouseBrainSlice'
-    processed_dir: './example_data/TestData/reconData/2018_10_02_MouseBrainSlice'
+    processed_dir: './example_data/TestData/reconData/generated'
     samples: ['SM_2018_1002_1633_1']
     positions: 'all'
     background: 'BG_2018_1002_1625_1'
