--- examples/example_configs/config_MouseBrainSlice2_workflow_test.yml
@@ -1,25 +1,0 @@
-dataset:
-    data_dir: './example_data/TestData/rawData/2018_10_02_MouseBrainSlice'
-    processed_dir: './example_data/TestData/reconData/2018_10_02_MouseBrainSlice'
-    samples: ['SM_2018_1002_1640_1']
-    positions: 'all'
-    background: 'BG_2018_1002_1625_1'
-#    BgDir_local: None
-
-processing:
-    output_channels: ['Brightfield_computed', 'Retardance', 'Orientation', 'Polarization']
-    circularity: 'rcp'
-#    bgCorrect: 'Local_filter'
-    background_correction: 'Input'
-    flatfield_correction: False
-#    batchProc: False
-    azimuth_offset: 0
-    separate_positions: True
-
-plotting:
-    normalize_color_images: True
-    save_birefringence_fig: False
-    save_stokes_fig: False
-    save_polarization_fig: False
-    save_micromanager_fig: False
-
